Extract export listings query
  * Extract export listings to it's own module.
  * Typo on exporters file path.
