+++ apps/re/lib/listings/listings.ex
@@ -204,14 +204,4 @@ defmodule Re.Listings do
     |> Queries.preload_relations(@full_preload)
     |> Repo.all()
   end
-
-  def exportable(%{state_slug: state_slug, city_slug: city_slug}, params) do
-    Queries.active()
-    |> Queries.by_state_slug_and_city_slug(state_slug, city_slug)
-    |> Queries.preload_relations(@partial_preload)
-    |> Queries.order_by_id()
-    |> Queries.offset(params)
-    |> Queries.limit(params)
-    |> Repo.all()
-  end
 end
